Benchmarking - c029: gfx_pc.cpp opcode-extraction comment hardening
This commit captures the documentation around the explicit uint8_t
cast in gfx_run_dl that prevents bit-window mis-truncation from a
64-bit uintptr_t w0. Without the mask, a garbage w0 like
0xfdbb0000ffff0000 would shift to 0x000000fdbb0000ff and the
implicit truncation to uint32_t produces a misleading opcode for
sysFatalError. The GBI opcode field is always a single byte (bits
24-31 of w0), so the cast is the correct extraction and the comment
locks the rationale in place against future "simplification."

Refs: c029

+++ port/fast3d/gfx_pc.cpp
@@ -2443,7 +2443,12 @@ static void gfx_run_dl(Gfx* cmd) {
     uint64_t ourHash = -1;
 
     for (;;) {
-        uint32_t opcode = cmd->words.w0 >> 24;
+        // GBI opcode is the high byte (bits 24-31) of w0. w0 is uintptr_t (64-bit
+        // on x86_64), so we must mask to 8 bits — otherwise a garbage w0 like
+        // 0xfdbb0000ffff0000 would shift to 0x000000fdbb0000ff and the implicit
+        // truncation to uint32_t produces a misleading opcode for sysFatalError.
+        // The actual GBI opcode field is always a single byte.
+        uint8_t opcode = (uint8_t)(cmd->words.w0 >> 24);
         // gfx_print_cmd(cmd);
         switch (opcode) {
                 // RSP commands:
